Split function expression families
- Move function-valued plan expression families into focused plan modules.
- Move runtime function expression evaluators into matching family modules.
- Keep typed return-family invariants intact before function-returning function support.
