+++ src/plan/expression/function.rs
@@ -1,39 +1,21 @@
-use super::{BoolExpr, IntExpr};
-use crate::plan::{
-    BoolFunctionLocalId, BoolFunctionValue, FunctionType, FunctionValue, FunctionValueKind,
-    IntFunctionLocalId, IntFunctionValue, NilFunctionLocalId, NilFunctionValue, Step,
-    StringFunctionLocalId, StringFunctionValue,
+mod bool;
+mod int;
+mod nil;
+mod string;
+
+use crate::plan::{FunctionType, FunctionValue, FunctionValueKind};
+
+pub use self::{
+    bool::BoolFunctionExpr, int::IntFunctionExpr, nil::NilFunctionExpr, string::StringFunctionExpr,
 };
-use ecow::EcoString;
-use num_bigint::BigInt;
+pub(crate) use self::{
+    bool::BoolFunctionExprKind, int::IntFunctionExprKind, nil::NilFunctionExprKind,
+    string::StringFunctionExprKind,
+};
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct FunctionExpr {
     kind: FunctionExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct IntFunctionExpr {
-    type_: FunctionType,
-    kind: IntFunctionExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct StringFunctionExpr {
-    type_: FunctionType,
-    kind: StringFunctionExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct BoolFunctionExpr {
-    type_: FunctionType,
-    kind: BoolFunctionExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct NilFunctionExpr {
-    type_: FunctionType,
-    kind: NilFunctionExprKind,
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -42,98 +24,6 @@ pub(crate) enum FunctionExprKind {
     String(StringFunctionExpr),
     Bool(BoolFunctionExpr),
     Nil(NilFunctionExpr),
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum IntFunctionExprKind {
-    Value(IntFunctionValue),
-    LocalGet {
-        local: IntFunctionLocalId,
-        name: EcoString,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<IntFunctionExpr>,
-        false_: Box<IntFunctionExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, IntFunctionExpr)>,
-        fallback: Box<IntFunctionExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<IntFunctionExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum StringFunctionExprKind {
-    Value(StringFunctionValue),
-    LocalGet {
-        local: StringFunctionLocalId,
-        name: EcoString,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<StringFunctionExpr>,
-        false_: Box<StringFunctionExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, StringFunctionExpr)>,
-        fallback: Box<StringFunctionExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<StringFunctionExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum BoolFunctionExprKind {
-    Value(BoolFunctionValue),
-    LocalGet {
-        local: BoolFunctionLocalId,
-        name: EcoString,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<BoolFunctionExpr>,
-        false_: Box<BoolFunctionExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, BoolFunctionExpr)>,
-        fallback: Box<BoolFunctionExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<BoolFunctionExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum NilFunctionExprKind {
-    Value(NilFunctionValue),
-    LocalGet {
-        local: NilFunctionLocalId,
-        name: EcoString,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<NilFunctionExpr>,
-        false_: Box<NilFunctionExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, NilFunctionExpr)>,
-        fallback: Box<NilFunctionExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<NilFunctionExpr>,
-    },
 }
 
 impl FunctionExpr {
@@ -218,278 +108,6 @@ impl FunctionExpr {
     }
 }
 
-impl IntFunctionExpr {
-    pub(crate) fn value(value: IntFunctionValue) -> Self {
-        Self {
-            type_: value.type_(),
-            kind: IntFunctionExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(
-        local: IntFunctionLocalId,
-        name: EcoString,
-        type_: FunctionType,
-    ) -> Self {
-        Self {
-            type_,
-            kind: IntFunctionExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn bool_case(
-        subject: BoolExpr,
-        true_: IntFunctionExpr,
-        false_: IntFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: true_.type_.clone(),
-            kind: IntFunctionExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, IntFunctionExpr)>,
-        fallback: IntFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: fallback.type_.clone(),
-            kind: IntFunctionExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: IntFunctionExpr) -> Self {
-        Self {
-            type_: return_.type_.clone(),
-            kind: IntFunctionExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub fn type_(&self) -> &FunctionType {
-        &self.type_
-    }
-
-    pub(crate) fn kind(&self) -> &IntFunctionExprKind {
-        &self.kind
-    }
-}
-
-impl StringFunctionExpr {
-    pub(crate) fn value(value: StringFunctionValue) -> Self {
-        Self {
-            type_: value.type_(),
-            kind: StringFunctionExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(
-        local: StringFunctionLocalId,
-        name: EcoString,
-        type_: FunctionType,
-    ) -> Self {
-        Self {
-            type_,
-            kind: StringFunctionExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn bool_case(
-        subject: BoolExpr,
-        true_: StringFunctionExpr,
-        false_: StringFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: true_.type_.clone(),
-            kind: StringFunctionExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, StringFunctionExpr)>,
-        fallback: StringFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: fallback.type_.clone(),
-            kind: StringFunctionExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: StringFunctionExpr) -> Self {
-        Self {
-            type_: return_.type_.clone(),
-            kind: StringFunctionExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub fn type_(&self) -> &FunctionType {
-        &self.type_
-    }
-
-    pub(crate) fn kind(&self) -> &StringFunctionExprKind {
-        &self.kind
-    }
-}
-
-impl BoolFunctionExpr {
-    pub(crate) fn value(value: BoolFunctionValue) -> Self {
-        Self {
-            type_: value.type_(),
-            kind: BoolFunctionExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(
-        local: BoolFunctionLocalId,
-        name: EcoString,
-        type_: FunctionType,
-    ) -> Self {
-        Self {
-            type_,
-            kind: BoolFunctionExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn bool_case(
-        subject: BoolExpr,
-        true_: BoolFunctionExpr,
-        false_: BoolFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: true_.type_.clone(),
-            kind: BoolFunctionExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, BoolFunctionExpr)>,
-        fallback: BoolFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: fallback.type_.clone(),
-            kind: BoolFunctionExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: BoolFunctionExpr) -> Self {
-        Self {
-            type_: return_.type_.clone(),
-            kind: BoolFunctionExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub fn type_(&self) -> &FunctionType {
-        &self.type_
-    }
-
-    pub(crate) fn kind(&self) -> &BoolFunctionExprKind {
-        &self.kind
-    }
-}
-
-impl NilFunctionExpr {
-    pub(crate) fn value(value: NilFunctionValue) -> Self {
-        Self {
-            type_: value.type_(),
-            kind: NilFunctionExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(
-        local: NilFunctionLocalId,
-        name: EcoString,
-        type_: FunctionType,
-    ) -> Self {
-        Self {
-            type_,
-            kind: NilFunctionExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn bool_case(
-        subject: BoolExpr,
-        true_: NilFunctionExpr,
-        false_: NilFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: true_.type_.clone(),
-            kind: NilFunctionExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, NilFunctionExpr)>,
-        fallback: NilFunctionExpr,
-    ) -> Self {
-        Self {
-            type_: fallback.type_.clone(),
-            kind: NilFunctionExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: NilFunctionExpr) -> Self {
-        Self {
-            type_: return_.type_.clone(),
-            kind: NilFunctionExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub fn type_(&self) -> &FunctionType {
-        &self.type_
-    }
-
-    pub(crate) fn kind(&self) -> &NilFunctionExprKind {
-        &self.kind
-    }
-}
-
 impl From<IntFunctionExpr> for FunctionExpr {
     fn from(expression: IntFunctionExpr) -> Self {
         Self::int(expression)
@@ -517,15 +135,13 @@ impl From<NilFunctionExpr> for FunctionExpr {
 #[cfg(test)]
 mod tests {
     use super::{
-        BoolFunctionExpr, BoolFunctionExprKind, FunctionExpr, FunctionExprKind, IntFunctionExpr,
-        IntFunctionExprKind, NilFunctionExpr, NilFunctionExprKind, StringFunctionExpr,
-        StringFunctionExprKind,
+        BoolFunctionExpr, FunctionExpr, FunctionExprKind, IntFunctionExpr, NilFunctionExpr,
+        StringFunctionExpr,
     };
     use crate::plan::{
-        BoolExpr, BoolFunctionLocalId, BoolFunctionValue, Expr, FunctionType, FunctionValue,
-        IntExpr, IntFunctionId, IntFunctionLocalId, IntFunctionValue, IntLocalId,
-        NilFunctionLocalId, NilFunctionValue, ParamLocal, RuntimeFunctionId, Step,
-        StringFunctionLocalId, StringFunctionValue, ValueType,
+        BoolFunctionId, BoolFunctionValue, FunctionValue, IntFunctionId, IntFunctionValue,
+        NilFunctionId, NilFunctionValue, ParamLocal, RuntimeFunctionId, StringFunctionId,
+        StringFunctionValue,
     };
 
     #[test]
@@ -593,178 +209,38 @@ mod tests {
         ));
     }
 
-    #[test]
-    fn int_function_expr_kind_accessors() {
-        assert_eq!(
-            int_function_type(),
-            FunctionType::new(vec![crate::plan::ValueType::Int], ValueType::Int),
-        );
-        assert!(matches!(
-            IntFunctionExpr::local_get(IntFunctionLocalId(0), "f".into(), int_function_type(),)
-                .kind(),
-            IntFunctionExprKind::LocalGet { .. }
-        ));
-        assert!(matches!(
-            IntFunctionExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), int_function_value())],
-                int_function_value(),
-            )
-            .kind(),
-            IntFunctionExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            IntFunctionExpr::block(
-                vec![Step::evaluate(Expr::int(IntExpr::value(1.into())))],
-                int_function_value(),
-            )
-            .kind(),
-            IntFunctionExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            IntFunctionExpr::bool_case(
-                BoolExpr::value(true),
-                int_function_value(),
-                int_function_value(),
-            )
-            .kind(),
-            IntFunctionExprKind::BoolCase { .. }
-        ));
-    }
-
-    #[test]
-    fn string_bool_nil_function_expr_kind_accessors() {
-        assert!(matches!(
-            StringFunctionExpr::local_get(
-                StringFunctionLocalId(0),
-                "f".into(),
-                string_function_value().type_().clone(),
-            )
-            .kind(),
-            StringFunctionExprKind::LocalGet { .. }
-        ));
-        assert!(matches!(
-            StringFunctionExpr::bool_case(
-                BoolExpr::value(true),
-                string_function_value(),
-                string_function_value(),
-            )
-            .kind(),
-            StringFunctionExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            StringFunctionExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), string_function_value())],
-                string_function_value(),
-            )
-            .kind(),
-            StringFunctionExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            StringFunctionExpr::block(Vec::new(), string_function_value()).kind(),
-            StringFunctionExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            BoolFunctionExpr::local_get(
-                BoolFunctionLocalId(0),
-                "f".into(),
-                bool_function_value().type_().clone(),
-            )
-            .kind(),
-            BoolFunctionExprKind::LocalGet { .. }
-        ));
-        assert!(matches!(
-            BoolFunctionExpr::bool_case(
-                BoolExpr::value(true),
-                bool_function_value(),
-                bool_function_value(),
-            )
-            .kind(),
-            BoolFunctionExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            BoolFunctionExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), bool_function_value())],
-                bool_function_value(),
-            )
-            .kind(),
-            BoolFunctionExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            BoolFunctionExpr::block(Vec::new(), bool_function_value()).kind(),
-            BoolFunctionExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            NilFunctionExpr::local_get(
-                NilFunctionLocalId(0),
-                "f".into(),
-                nil_function_value().type_().clone(),
-            )
-            .kind(),
-            NilFunctionExprKind::LocalGet { .. }
-        ));
-        assert!(matches!(
-            NilFunctionExpr::bool_case(
-                BoolExpr::value(true),
-                nil_function_value(),
-                nil_function_value(),
-            )
-            .kind(),
-            NilFunctionExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            NilFunctionExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), nil_function_value())],
-                nil_function_value(),
-            )
-            .kind(),
-            NilFunctionExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            NilFunctionExpr::block(Vec::new(), nil_function_value()).kind(),
-            NilFunctionExprKind::Block { .. }
-        ));
-    }
-
     fn function_value() -> FunctionValue {
-        FunctionValue::new(RuntimeFunctionId::Int(IntFunctionId(0)), vec![int_param(0)])
+        FunctionValue::new(
+            RuntimeFunctionId::Int(IntFunctionId(0)),
+            vec![ParamLocal::int(crate::plan::IntLocalId(0))],
+        )
     }
 
     fn int_function_value() -> IntFunctionExpr {
-        IntFunctionExpr::value(IntFunctionValue::new(IntFunctionId(0), vec![int_param(0)]))
+        IntFunctionExpr::value(IntFunctionValue::new(
+            IntFunctionId(0),
+            vec![ParamLocal::int(crate::plan::IntLocalId(0))],
+        ))
     }
 
     fn string_function_value() -> StringFunctionExpr {
         StringFunctionExpr::value(StringFunctionValue::new(
-            crate::plan::StringFunctionId(0),
-            vec![crate::plan::ParamLocal::string(crate::plan::StringLocalId(
-                0,
-            ))],
+            StringFunctionId(0),
+            vec![ParamLocal::string(crate::plan::StringLocalId(0))],
         ))
     }
 
     fn bool_function_value() -> BoolFunctionExpr {
         BoolFunctionExpr::value(BoolFunctionValue::new(
-            crate::plan::BoolFunctionId(0),
-            vec![crate::plan::ParamLocal::bool(crate::plan::BoolLocalId(0))],
+            BoolFunctionId(0),
+            vec![ParamLocal::bool(crate::plan::BoolLocalId(0))],
         ))
     }
 
     fn nil_function_value() -> NilFunctionExpr {
         NilFunctionExpr::value(NilFunctionValue::new(
-            crate::plan::NilFunctionId(0),
-            vec![crate::plan::ParamLocal::nil(crate::plan::NilLocalId(0))],
+            NilFunctionId(0),
+            vec![ParamLocal::nil(crate::plan::NilLocalId(0))],
         ))
-    }
-
-    fn int_function_type() -> FunctionType {
-        FunctionType::new(vec![crate::plan::ValueType::Int], ValueType::Int)
-    }
-
-    fn int_param(index: usize) -> ParamLocal {
-        ParamLocal::int(IntLocalId(index))
     }
 }
